+++ src/hepattn/experiments/trackml/configs/filtering.yaml
