+++ src/hepattn/experiments/trackml/configs/tracking.yaml
