+++ src/hepattn/experiments/trackml/main.py
@@ -3,15 +3,15 @@ import comet_ml
 
 from lightning.pytorch.cli import ArgsType
 from hepattn.utils.cli import CLI
-from hepattn.experiments.tracking.data import TrackingDataModule
-from hepattn.experiments.tracking.tracker import Tracker
+from hepattn.experiments.trackml.data import TrackMLDataModule
+from hepattn.experiments.trackml.model import TrackMLTracker
 
 
 
 def main(args: ArgsType = None) -> None:
     CLI(
-        model_class=Tracker,
-        datamodule_class=TrackingDataModule,
+        model_class=TrackMLTracker,
+        datamodule_class=TrackMLDataModule,
         args=args,
         parser_kwargs={"default_env": True},
     )
